modify: house 리뷰 파일 경로 수정
